Move type_aware_deep_merge() into ValueExt
